+++ provision/core/git/provision.sh
@@ -4,8 +4,11 @@ set -eo pipefail
 
 # @noargs
 function git_register_apt_sources() {
-  local OSID=$(lsb_release --id --short)
-  local OSCODENAME=$(lsb_release --codename --short)
+  local OSCODENAME
+  local OSID
+
+  OSCODENAME=$(lsb_release --codename --short)
+  OSID=$(lsb_release --id --short)
 
   if [ "${OSID}" == "Ubuntu" ]; then
     if ! vvv_src_list_has "git-core/ppa"; then
@@ -39,11 +42,8 @@ vvv_add_hook register_apt_packages git_register_apt_packages
 
 # @noargs
 function git_register_apt_keys() {
-  if ! vvv_apt_keys_has 'git-lfs'; then
-    # Apply the PackageCloud signing key which signs git lfs
-    vvv_info " * Applying the PackageCloud Git-LFS signing key..."
-    apt-key add /srv/provision/core/git/apt-keys/git-lfs.key
-  fi
+  cp -f /srv/provision/core/git/apt-keys/github_git-lfs-archive-keyring.gpg /etc/apt/keyrings/github_git-lfs-archive-keyring.gpg
+  chmod 0644 /etc/apt/keyrings/github_git-lfs-archive-keyring.gpg
 }
 vvv_add_hook register_apt_keys git_register_apt_keys
 
@@ -51,15 +51,15 @@ vvv_add_hook register_apt_keys git_register_apt_keys
 function git_after_packages() {
   # if this setting isn't set, git will exit and provisioning will fail
   if ! git config --global pull.rebase; then
-    vvv_info " * Git hasn't been told how to merge branches, setting pull.rebase false for the merge strategy"
-    git config --global pull.rebase false
+    vvv_info " * Git global config hasn't been told how to merge branches, setting pull.rebase false for the merge strategy"
+    git config --global pull.rebase false || vvv_warn "Failed ot update git config pull.rebase to false"
   fi
   if ! noroot git config --global pull.rebase; then
-    vvv_info " * Git hasn't been told how to merge branches, setting pull.rebase false for the merge strategy"
-    noroot git config --global pull.rebase false
+    vvv_info " * Git noroot global config hasn't been told how to merge branches, setting pull.rebase false for the merge strategy"
+    noroot git config --global pull.rebase false || vvv_warn "Failed to update noroot git config pull.rebase to false"
   fi
 
-  git config --global --add safe.directory '*'
-  noroot git config --global --add safe.directory '*'
+  git config --global --add safe.directory '*' || vvv_warn "Failed to update git config safe.directory"
+  noroot git config --global --add safe.directory '*'|| vvv_warn "Failed to update noroot git config safe.directory"
 }
 vvv_add_hook after_packages git_after_packages