+++ provision/core/php/provision.sh
@@ -153,17 +153,16 @@ vvv_add_hook nginx_upstreams php_nginx_upstream
 
 function vvv_php_memcached_register_packages() {
   VVV_PACKAGE_LIST+=(
-    php${VVV_BASE_PHPVERSION}-memcache
-    php${VVV_BASE_PHPVERSION}-memcached
+    "php${VVV_BASE_PHPVERSION}-memcache"
+    "php${VVV_BASE_PHPVERSION}-memcached"
   )
 }
 export -f vvv_php_memcached_register_packages
 vvv_add_hook before_packages vvv_php_memcached_register_packages
 
-
 function vvv_php_redis_register_packages() {
   VVV_PACKAGE_LIST+=(
-    php${VVV_BASE_PHPVERSION}-redis
+    "php${VVV_BASE_PHPVERSION}-redis"
   )
 }
 export -f vvv_php_redis_register_packages